+++ MatchRetriever/ModelFactories/MatchSelectionModelFactory.cs
@@ -1,4 +1,6 @@
-﻿using MatchRetriever.Models;
+﻿using MatchRetriever.Configuration;
+using MatchRetriever.Enumerals;
+using MatchRetriever.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,13 +10,18 @@ namespace MatchRetriever.ModelFactories
 {
     public interface IMatchSelectionModelFactory
     {
-        Task<MatchSelectionModel> GetModel(long steamId, int dailyLimit);
+        Task<MatchSelectionModel> GetModel(long steamId, SubscriptionType subscriptionType);
     }
 
     public class MatchSelectionModelFactory : ModelFactoryBase, IMatchSelectionModelFactory
     {
-        public MatchSelectionModelFactory(IServiceProvider sp) : base(sp)
+        public ISubscriptionConfigProvider _subscriptionConfigLoader;
+
+        public MatchSelectionModelFactory(
+            IServiceProvider sp,
+            ISubscriptionConfigProvider subscriptionConfigLoader) : base(sp)
         {
+            _subscriptionConfigLoader = subscriptionConfigLoader;
         }
 
         /// <summary>
@@ -22,11 +29,22 @@ namespace MatchRetriever.ModelFactories
         /// Users are allowed `dailyLimit` matches for each day, with the reset occuring at 00:00 UTC.
         /// </summary>
         /// <param name="steamId"></param>
-        /// <param name="dailyLimit"></param>
+        /// <param name="subscriptionType"></param>
         /// <returns></returns>
-        public async Task<MatchSelectionModel> GetModel(long steamId, int dailyLimit)
+        public async Task<MatchSelectionModel> GetModel(long steamId, SubscriptionType subscriptionType)
         {
-            var allMatches = _context.PlayerMatchStats.Where(x => x.SteamId == steamId)
+            MatchSelectionModel matchSelectionModel = new MatchSelectionModel
+            {
+                Matches = null,
+                DailyLimitReached = false,
+                DailyLimitEnds = DateTime.MaxValue,
+                InaccessibleBefore = DateTime.MinValue,
+            };
+
+            var config = _subscriptionConfigLoader.Config.SettingsFromSubscriptionType(subscriptionType);
+
+            #region Select All Matches
+            var matches = _context.PlayerMatchStats.Where(x => x.SteamId == steamId)
                 .Select(x => new MatchSelectionModel.Match
                 {
                     MatchId = x.MatchId,
@@ -36,27 +54,68 @@ namespace MatchRetriever.ModelFactories
                 })
                 .ToList();
 
-            // apply the daily limit
-            var allowedMatches = allMatches.GroupBy(x => x.MatchDate.ToUniversalTime().DayOfYear)
+            #endregion
+
+            # region Apply Inaccessible Limit
+            // If the value of MatchAccessDurationInDays is NOI -1, apply the limit.
+            if (config.MatchAccessDurationInDays != -1)
+            {
+                // Subtract the matchAccessDurationInDays from NOW.
+                matchSelectionModel.InaccessibleBefore = DateTime.Now.Subtract(
+                    TimeSpan.FromDays(config.MatchAccessDurationInDays));
+
+                matches = ApplyInaccessibleLimit(matches, matchSelectionModel.InaccessibleBefore);
+            }
+            #endregion
+
+            #region Apply Daily Limit
+            // If the value of DailyMatchesLimit is NOI -1, apply the limit.
+            if (config.DailyMatchesLimit != -1)
+            {
+                matches = ApplyDailyLimit(matches, config.DailyMatchesLimit);
+
+                matchSelectionModel.DailyLimitReached = matches
+                    .Where(x => x.MatchDate.ToUniversalTime().DayOfYear == DateTime.Now.ToUniversalTime().DayOfYear)
+                    .Count() >= config.DailyMatchesLimit;
+
+                matchSelectionModel.DailyLimitEnds = DateTime.Now.AddDays(1).ToUniversalTime();
+            }
+            #endregion
+
+            matchSelectionModel.Matches = matches;
+            return matchSelectionModel;
+
+
+        }
+
+        /// <summary>
+        /// Returns only matches that happen after inaccessibleBefore.
+        /// </summary>
+        /// <param name="matches">Matches to limit</param>
+        /// <param name="inaccessibleBefore"></param>
+        /// <returns></returns>
+        public List<MatchSelectionModel.Match> ApplyInaccessibleLimit(List<MatchSelectionModel.Match> matches, DateTime inaccessibleBefore)
+        {
+            return matches
+                .Where(x => x.MatchDate >= inaccessibleBefore)
+                .ToList();
+        }
+
+        /// <summary>
+        /// Apply a specified DailyLimit on a list of Matches.
+        /// </summary>
+        /// <param name="matches">Matches to limit</param>
+        /// <param name="dailyLimit"></param>
+        /// <returns></returns>
+        public List<MatchSelectionModel.Match> ApplyDailyLimit(List<MatchSelectionModel.Match> matches, int dailyLimit)
+        {
+            return matches.GroupBy(x => x.MatchDate.ToUniversalTime().DayOfYear)
                 .OrderByDescending(x => x.Key)
                 .SelectMany(x => x
                     .OrderBy(y => y.MatchDate)
                     .Take(dailyLimit))
-                .OrderBy(x=>x.MatchDate)
+                .OrderBy(x => x.MatchDate)
                 .ToList();
-
-            var dailyLimitReached = allMatches
-                .Where(x => x.MatchDate.ToUniversalTime().DayOfYear == DateTime.Now.ToUniversalTime().DayOfYear)
-                .Count() > dailyLimit;
-
-            DateTime dailyLimitEnds = DateTime.Now.AddDays(1).ToUniversalTime();
-
-            return new MatchSelectionModel
-            {
-                Matches = allowedMatches,
-                DailyLimitReached = dailyLimitReached,
-                DailyLimitEnds = dailyLimitEnds,
-            };
         }
     }
 }