+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-04-30
+### Added
+- SubscriptionConfiguration from JSON
+    - Loads `MatchAccessDurationInDays`, `DailyMatchesLimit` and `PositionFramesPerSecond` from a file on disk,
+    - Applies these limits to the `MatchSelection` model factory.
+- Environment Variables
+    - `MOCK_STEAM_USER_OPERATOR`
+    - `MOCK_SUBSCRIPTION_LOADER`
+
+### Changed
+- `MatchSelection` Controller now expects a `SubscriptionType` as a replacement for the `dailyLimit`.
+- Default mock behavior of SteamUserOperator - Now requires an ENV Var
+
 ## [1.3.0] - 2020-04-22
 ### Added
 - Change MatchSelection algorithm back to 24h limit resetting at midnight UTC